Change how wxWidgets is added
wxWidgets is now added as its own repository instead of as a submodule for monolithic builds. This makes it much easier to make commits, as one does not have to be careful anymore to not commit the submodule by mistake, as it is ignored completley with the gitignore file. This ensures REUSE compliance.