Move Input stuff to DesktopModel
This will make it testable.
The timer is now stored in a class that overrides an abstract base, which is passed into the model functions that need to update the timer.
DestkopModel is a new class that inherits from GameModel, which extends the GameModel with desktop input functionality.
This is going to break all of the tests. They need to be remade next step.